Keep Scrollable.kt with less changes comparing to AOSP
- remove refactorings. Usually they are good, but not in case we need to constantly merge the upstream
- move the new functionality to the new file. Otherwise, it can conflict with near-code changes, and pollutes the merge resolver diff
